Move feature test to new project
Fix codecov output error

Ignore test folder
