+++ codecov.yml
@@ -19,4 +19,8 @@ flags:
   featuretest:
     # only test featuretest
     paths:
-      - aliyun-net-feature-test/
+      - aliyun-sdk-feature-test/
+
+ignore:
+  - "aliyun-net-sdk-core.Tests/"  # ignore aliyun-net-sdk-core.Tests
+  - "aliyun-sdk-feature-test/"    # ignore aliyun-sdk-feature-test
